+++ resources/web/docs_js/components/console_widget.jsx
@@ -5,6 +5,7 @@ import linkState from "../../../../../node_modules/linkstate";
 import {connect} from "../../../../../node_modules/preact-redux";
 import {openModal} from "../actions/modal";
 import {saveSettings} from "../actions/settings";
+import AlternativePicker from "./alternative_picker";
 
 const copyAsCurl = ({setting, consoleText, isKibana}) => (_, getState) => {
   const state       = getState();
@@ -68,68 +69,6 @@ export const ConsoleForm = connect((state, props) =>
         `${props.setting}_curl_user`,
         `${props.setting}_curl_password`], state.settings)
 , {saveSettings})(_ConsoleForm);
-
-const alternativePrettyName = rawName => {
-  switch(rawName) {
-    case 'console': return 'Console';
-    case 'csharp': return 'C#';
-    case 'js': return 'JavaScript';
-    case 'php': return 'PHP';
-    default: return rawName;
-  }
-};
-
-const AlternativeChoice = ({name: name}) => {
-  return <option value={name}>{alternativePrettyName(name)}</option>;
-};
-
-export const _AlternativePicker = ({
-  alternatives: alternatives,
-  consoleAlternative: consoleAlternative,
-  saveSettings: saveSettings,
-}) => {
-  if (!alternatives) {
-    return <div/>; // Empty div to preserve the spacing
-  }
-  const consoleAlternatives = alternatives.console;
-  if (!consoleAlternatives) {
-    return <div/>;
-  }
-
-  const items = [];
-  let sawChoice = 'console' === consoleAlternative;
-  items.push(<AlternativeChoice name='console'/>);
-  for (const name of Object.keys(consoleAlternatives)) {
-    sawChoice |= name === consoleAlternative;
-    items.push(<AlternativeChoice name={name} />);
-  }
-
-  /* If value isn't in the list then *make* it and we'll render our standard
-   * "there no example for this language" option. This prevents us from
-   * squashing preferences that users set. */
-  if (!sawChoice) {
-    items.push(<AlternativeChoice name={consoleAlternative} />);
-  }
-  // TODO we shouldn't change these drop downs after the first time they are rendered. The extra choice should stay while you stay on the page. Maybe we can get away with rendering this once on page load and never subscribing again?
-
-  // TODO add the "message" bubble to the warning.
-  // TODO prevent "jumping" when the size of the snippets isn't the same
-  return <div className="AlternativePicker u-space-between">
-    <select className="AlternativePicker-select"
-            value={consoleAlternative}
-            onChange={(e) => saveSettings({consoleAlternative: e.target.value})}>
-      {items}
-    </select>
-    <div className="AlternativePicker-warning" />
-  </div>;
-};
-
-// TODO move me to my own file
-const AlternativePicker = connect(
-  (state, props) => merge(
-    pick(["consoleAlternative"], state.settings),
-    {alternatives: state.alternatives}),
-  {saveSettings})(_AlternativePicker)
 
 // ConsoleWidget isn't quite the right name for this any more....
 export const ConsoleWidget = props => {
